+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 
 - `custom_html_element!` no longer has a `custom_events` section. You can now just use `elements::CustomEvent` in the `events` section. See `examples/web-components-wrapper`.
+- `StrAttribute` derive macro to `impl Attribute` for types that implement `AsRef<str>`
 
 ## 0.9.0 - 2025-01-14